+++ src/images/cropped_image.rs
@@ -1,73 +1,37 @@
-use crate::{CropBoxError, ImageView, ImageViewMut};
+use crate::images::{check_crop_box, TypedCroppedImage, TypedCroppedImageMut};
+use crate::{
+    CropBoxError, ImageView, ImageViewMut, IntoImageView, IntoImageViewMut, PixelTrait, PixelType,
+};
 
-fn check_crop_box(
-    image_view: &impl ImageView,
-    left: u32,
-    top: u32,
-    width: u32,
-    height: u32,
-) -> Result<(), CropBoxError> {
-    let img_width = image_view.width();
-    let img_height = image_view.height();
-
-    if left >= img_width || top >= img_height {
-        return Err(CropBoxError::PositionIsOutOfImageBoundaries);
-    }
-    let right = left + width;
-    let bottom = top + height;
-    if right > img_width || bottom > img_height {
-        return Err(CropBoxError::SizeIsOutOfImageBoundaries);
-    }
-    Ok(())
-}
-
-macro_rules! image_view_impl {
-    ($wrapper_name:ident<$view_trait:ident>) => {
-        unsafe impl<'a, V: $view_trait> ImageView for $wrapper_name<'a, V> {
-            type Pixel = V::Pixel;
-
-            fn width(&self) -> u32 {
-                self.width
-            }
-
-            fn height(&self) -> u32 {
-                self.height
-            }
-
-            fn iter_rows(&self, start_row: u32) -> impl Iterator<Item = &[Self::Pixel]> {
-                let left = self.left as usize;
-                let right = left + self.width as usize;
-                self.image_view
-                    .iter_rows(self.top + start_row)
-                    .take((self.height - start_row) as usize)
-                    // SAFETY: correct values of the left and the right
-                    // are guaranteed by new() method.
-                    .map(move |row| unsafe { row.get_unchecked(left..right) })
-            }
-        }
-    };
-}
-
-/// It is a typed wrapper that provides [ImageView] for part of wrapped image.
-pub struct TypedCroppedImage<'a, V: ImageView> {
-    image_view: &'a V,
+/// It is a wrapper that provides [IntoImageView] for part of wrapped image.
+pub struct CroppedImage<'a, V: IntoImageView> {
+    image: &'a V,
     left: u32,
     top: u32,
     width: u32,
     height: u32,
 }
 
-impl<'a, V: ImageView> TypedCroppedImage<'a, V> {
+/// It is a wrapper that provides [IntoImageView] and [IntoImageViewMut] for part of wrapped image.
+pub struct CroppedImageMut<'a, V: IntoImageView> {
+    image: &'a mut V,
+    left: u32,
+    top: u32,
+    width: u32,
+    height: u32,
+}
+
+impl<'a, V: IntoImageView> CroppedImage<'a, V> {
     pub fn new(
-        image_view: &'a V,
+        image: &'a V,
         left: u32,
         top: u32,
         width: u32,
         height: u32,
     ) -> Result<Self, CropBoxError> {
-        check_crop_box(image_view, left, top, width, height)?;
+        check_crop_box(image.width(), image.height(), left, top, width, height)?;
         Ok(Self {
-            image_view,
+            image,
             left,
             top,
             width,
@@ -76,27 +40,17 @@ impl<'a, V: ImageView> TypedCroppedImage<'a, V> {
     }
 }
 
-image_view_impl!(TypedCroppedImage<ImageView>);
-
-/// It is a typed wrapper that provides [ImageView] and [ImageViewMut] for part of wrapped image.
-pub struct TypedCroppedImageMut<'a, V: ImageViewMut> {
-    image_view: &'a mut V,
-    left: u32,
-    top: u32,
-    width: u32,
-    height: u32,
-}
-impl<'a, V: ImageViewMut> TypedCroppedImageMut<'a, V> {
+impl<'a, V: IntoImageView> CroppedImageMut<'a, V> {
     pub fn new(
-        image_view: &'a mut V,
+        image: &'a mut V,
         left: u32,
         top: u32,
         width: u32,
         height: u32,
     ) -> Result<Self, CropBoxError> {
-        check_crop_box(image_view, left, top, width, height)?;
+        check_crop_box(image.width(), image.height(), left, top, width, height)?;
         Ok(Self {
-            image_view,
+            image,
             left,
             top,
             width,
@@ -105,17 +59,50 @@ impl<'a, V: ImageViewMut> TypedCroppedImageMut<'a, V> {
     }
 }
 
-image_view_impl!(TypedCroppedImageMut<ImageViewMut>);
+impl<'a, V: IntoImageView> IntoImageView for CroppedImage<'a, V> {
+    fn pixel_type(&self) -> Option<PixelType> {
+        self.image.pixel_type()
+    }
 
-unsafe impl<'a, V: ImageViewMut> ImageViewMut for TypedCroppedImageMut<'a, V> {
-    fn iter_rows_mut(&mut self, start_row: u32) -> impl Iterator<Item = &mut [Self::Pixel]> {
-        let left = self.left as usize;
-        let right = left + self.width as usize;
-        self.image_view
-            .iter_rows_mut(self.top + start_row)
-            .take((self.height - start_row) as usize)
-            // SAFETY: correct values of the left and the right
-            // are guaranteed by new() method.
-            .map(move |row| unsafe { row.get_unchecked_mut(left..right) })
+    fn width(&self) -> u32 {
+        self.width
+    }
+
+    fn height(&self) -> u32 {
+        self.height
+    }
+
+    fn image_view<P: PixelTrait>(&self) -> Option<impl ImageView<Pixel = P>> {
+        self.image.image_view().map(|v| {
+            TypedCroppedImage::new(v, self.left, self.top, self.width, self.height).unwrap()
+        })
+    }
+}
+
+impl<'a, V: IntoImageView> IntoImageView for CroppedImageMut<'a, V> {
+    fn pixel_type(&self) -> Option<PixelType> {
+        self.image.pixel_type()
+    }
+
+    fn width(&self) -> u32 {
+        self.width
+    }
+
+    fn height(&self) -> u32 {
+        self.height
+    }
+
+    fn image_view<P: PixelTrait>(&self) -> Option<impl ImageView<Pixel = P>> {
+        self.image.image_view().map(|v| {
+            TypedCroppedImage::new(v, self.left, self.top, self.width, self.height).unwrap()
+        })
+    }
+}
+
+impl<'a, V: IntoImageViewMut> IntoImageViewMut for CroppedImageMut<'a, V> {
+    fn image_view_mut<P: PixelTrait>(&mut self) -> Option<impl ImageViewMut<Pixel = P>> {
+        self.image.image_view_mut().map(|v| {
+            TypedCroppedImageMut::new(v, self.left, self.top, self.width, self.height).unwrap()
+        })
     }
 }